spearbit/96 (#248)
* fix return types as per https://github.com/spearbit-audits/astaria/issues/96

* fix return types

+++ src/LienToken.sol
@@ -704,7 +704,7 @@ contract LienToken is ERC721, ILienToken, Auth {
     }
   }
 
-  function getOwed(Stack memory stack) external view returns (uint192) {
+  function getOwed(Stack memory stack) external view returns (uint88) {
     validateLien(stack.lien);
     return _getOwed(stack, block.timestamp);
   }
@@ -712,7 +712,7 @@ contract LienToken is ERC721, ILienToken, Auth {
   function getOwed(Stack memory stack, uint256 timestamp)
     external
     view
-    returns (uint192)
+    returns (uint88)
   {
     validateLien(stack.lien);
     return _getOwed(stack, timestamp);
